+++ docs/release_notes.md
@@ -17,7 +17,7 @@
 ### 📈 Improvements
 
 - [Orchestration] Added new API `OrchestrationTemplateReference#withScope` to support prompt templates with resource-group scope.
-- [Orchestration] Chat completion calls now can have multiple module configs to support fallback modules.
+- [Orchestration] Chat completion calls now can have multiple module configs to support [fallback modules](https://sap.github.io/ai-sdk/docs/java/orchestration/chat-completion).
 
 ### 🐛 Fixed Issues